Add documentation to exception sample and rename class
